Fix helpers
