+++ test/support/helpers/review-charge-reference.helper.js
@@ -1,28 +1,29 @@
 'use strict'
 
 /**
- * @module ReviewChargeVersionsHelper
+ * @module ReviewChargeReferenceHelper
  */
 
 const { generateUUID } = require('../../../app/lib/general.lib.js')
-const ReviewChargeElementsReturnsModel = require('../../../app/models/review-charge-versions.model.js')
+const ReviewChargeReferenceModel = require('../../../app/models/review-charge-reference.model.js')
 
 /**
- * Add a new review charge elements returns record for 2pt matching
+ * Add a new review charge version record for 2pt matching
  *
  * If no `data` is provided, default values will be used. These are
  *
- * - `reviewChargeElementId` - [random UUID]
- * - `reviewReturnId` - [random UUID]
+ * - `reviewChargeVersionId` - [random UUID]
+ * - `chargeReferenceId` - [random UUID]
+ * - `aggregate` - Strategic review of charges (SRoC)
  *
  * @param {Object} [data] Any data you want to use instead of the defaults used here or in the database
  *
- * @returns {Promise<module:ReviewChargeVersionsModel>} The instance of the newly created record
+ * @returns {Promise<module:ReviewChargeReferenceModel>} The instance of the newly created record
  */
 function add (data = {}) {
   const insertData = defaults(data)
 
-  return ReviewChargeElementsReturnsModel.query()
+  return ReviewChargeReferenceModel.query()
     .insert({ ...insertData })
     .returning('*')
 }
@@ -37,8 +38,9 @@ function add (data = {}) {
  */
 function defaults (data = {}) {
   const defaults = {
-    reviewChargeElementId: generateUUID(),
-    reviewReturnId: generateUUID()
+    reviewChargeVersionId: generateUUID(),
+    chargeReference: generateUUID(),
+    aggregate: 1
   }
 
   return {
