+++ test/support/helpers/review-charge-version.helper.js
@@ -1,11 +1,11 @@
 'use strict'
 
 /**
- * @module ReviewChargeVersionsHelper
+ * @module ReviewChargeVersionHelper
  */
 
 const { generateUUID } = require('../../../app/lib/general.lib.js')
-const ReviewChargeVersionsModel = require('../../../app/models/review-charge-versions.model.js')
+const ReviewChargeVersionModel = require('../../../app/models/review-charge-version.model.js')
 
 /**
  * Add a new review charge version record for 2pt matching
@@ -20,12 +20,12 @@ const ReviewChargeVersionsModel = require('../../../app/models/review-charge-ver
  *
  * @param {Object} [data] Any data you want to use instead of the defaults used here or in the database
  *
- * @returns {Promise<module:ReviewChargeVersionsModel>} The instance of the newly created record
+ * @returns {Promise<module:ReviewChargeVersionModel>} The instance of the newly created record
  */
 function add (data = {}) {
   const insertData = defaults(data)
 
-  return ReviewChargeVersionsModel.query()
+  return ReviewChargeVersionModel.query()
     .insert({ ...insertData })
     .returning('*')
 }
